Move Events to Likeable namespace
