+++ src/Like/Observers/LikeObserver.php
@@ -14,10 +14,10 @@ namespace Cog\Laravel\Likeable\Like\Observers;
 use Cog\Contracts\Likeable\Like\Models\Like as LikeContract;
 use Cog\Contracts\Likeable\Services\LikeableService as LikeableServiceContract;
 use Cog\Laravel\Likeable\Like\Enums\LikeType;
-use Cog\Laravel\Likeable\Like\Events\ModelWasDisliked;
-use Cog\Laravel\Likeable\Like\Events\ModelWasLiked;
-use Cog\Laravel\Likeable\Like\Events\ModelWasUndisliked;
-use Cog\Laravel\Likeable\Like\Events\ModelWasUnliked;
+use Cog\Laravel\Likeable\Likeable\Events\ModelWasDisliked;
+use Cog\Laravel\Likeable\Likeable\Events\ModelWasLiked;
+use Cog\Laravel\Likeable\Likeable\Events\ModelWasUndisliked;
+use Cog\Laravel\Likeable\Likeable\Events\ModelWasUnliked;
 
 /**
  * Class LikeObserver.
