+++ tests/Unit/Likeable/Events/ModelWasDislikedTest.php
@@ -9,28 +9,28 @@
  * file that was distributed with this source code.
  */
 
-namespace Cog\Tests\Laravel\Likeable\Unit\Like\Events;
+namespace Cog\Tests\Laravel\Likeable\Unit\Likeable\Events;
 
-use Cog\Laravel\Likeable\Like\Events\ModelWasLiked;
+use Cog\Laravel\Likeable\Likeable\Events\ModelWasDisliked;
 use Cog\Tests\Laravel\Likeable\Stubs\Models\Entity;
 use Cog\Tests\Laravel\Likeable\TestCase;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 
 /**
- * Class ModelWasLikedTest.
+ * Class ModelWasDislikedTest.
  *
- * @package Cog\Tests\Laravel\Likeable\Unit\Like\Events
+ * @package Cog\Tests\Laravel\Likeable\Unit\Likeable\Events
  */
-class ModelWasLikedTest extends TestCase
+class ModelWasDislikedTest extends TestCase
 {
     use DatabaseTransactions;
 
     /** @test */
     public function it_can_fire_model_was_liked_event()
     {
-        $this->expectsEvents(ModelWasLiked::class);
+        $this->expectsEvents(ModelWasDisliked::class);
 
         $entity = factory(Entity::class)->create();
-        $entity->like(1);
+        $entity->dislike(1);
     }
 }
