+++ tests/Unit/Likeable/Events/ModelWasUndislikedTest.php
@@ -9,9 +9,9 @@
  * file that was distributed with this source code.
  */
 
-namespace Cog\Tests\Laravel\Likeable\Unit\Like\Events;
+namespace Cog\Tests\Laravel\Likeable\Unit\Likeable\Events;
 
-use Cog\Laravel\Likeable\Like\Events\ModelWasUndisliked;
+use Cog\Laravel\Likeable\Likeable\Events\ModelWasUndisliked;
 use Cog\Tests\Laravel\Likeable\Stubs\Models\Entity;
 use Cog\Tests\Laravel\Likeable\TestCase;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
@@ -19,7 +19,7 @@ use Illuminate\Foundation\Testing\DatabaseTransactions;
 /**
  * Class ModelWasUndislikedTest.
  *
- * @package Cog\Tests\Laravel\Likeable\Unit\Like\Events
+ * @package Cog\Tests\Laravel\Likeable\Unit\Likeable\Events
  */
 class ModelWasUndislikedTest extends TestCase
 {
